% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find leap.R
\name{find_leap}
\alias{find_leap}
\title{Leap Day Finder}
\usage{
find_leap(x)
}
\arguments{
\item{x}{Date or Date Vectors}
}
\value{
Logical Vector (Leap Days are TRUE)
}
\description{
Finds any leap days in a vector of dates.

The Input should be a vector of dates that has at least the month and day attached to it. find_leap returns a logical vector that displays leap days as true.
}
\examples{

dates<-c("date1","date2","date3")
date<-find_leap(dates)

}
\seealso{
\code{\link{lubridate}}
}
