% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{intensity}
\alias{intensity}
\title{Intensity Value}
\usage{
intensity(x, ind=dated)
}
\arguments{
\item{x}{RasterBrick or TimeSeries}

\item{ind}{Index of Dates from the lpdates function (should be saved as 'dated')}
}
\value{
RasterBrick or TimeSeries of Yearly data
}
\description{
Calculate the intensity of the mid summer drought from a RasterBrick or a Time Series, or the average of the two onset data points minus the minimum data point.

The input must be in the form of daily data, with the first data point being January 1st of a respective year.

If x is a RasterBrick, then the output is a RasterBrick with a data point for each year. If you would like to find a statistical value of the new rasterbrick, it is recommended to use

r <- raster::calc(x, mean)
}
\examples{
# using RasterBricks
r<-raster::calc(x, intensity)

# using TimeSeries
r<-intensity(x)
}
\seealso{
\code{\link{raster}}
}
