% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulling dates.R
\name{lpdates}
\alias{lpdates}
\title{Pulling Dates}
\usage{
lpdates(length=20440, date_start="1950-01-01", cal="standard", date1="05-01", date2="07-15", date3="07-16", date4="09-30")
}
\arguments{
\item{cal}{"standard" corresponds to a data set that has leap days. "365_day" corresponds to a data set that has 365 days. No matter the input, the function will correct for it.}

\item{date1}{Start index to find first onset date}

\item{date2}{End index to find first onset date}

\item{date3}{Start index to find second onset date}

\item{date4}{End index to find second onset date}

\item{x}{RasterBrick or TimeSeries}

\item{start_date}{Start Date of the input. NOTE: This MUST be January 1st, but the year can differ.}
}
\value{
A vector of indexes that correspond to the four dates every year, with respect to the start date.

NOTE: It would be significantly easier to save the output as a variable called 'dated', especially when using rasters so that you do not have to create an intermediate function.
}
\description{
This function pulls the dates that correspond with the Mid Summer Drought, and is used in conjuction with the duration, minimum, onset, and intensity functions.

The main input is the length of the data you are working with, which is either the length of a vector or the number of layers of a raster.
}
\examples{
# using RasterBricks
dated<-dur(length=nlayers(r), date_start="1950-01-01", date1="05-01", date2="07-15", date3="07-16", date4="09-30")

# using TimeSeries
dated<-dur(length=length(t), ...)
}
