% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onset.R
\name{onset}
\alias{onset}
\title{Onset Value}
\usage{
onset(x, ind=dated)
}
\arguments{
\item{x}{RasterBrick or TimeSeries}

\item{ind}{Index of Dates from the lpdates function (should be saved as 'dated')}
}
\value{
RasterBrick or TimeSeries of Yearly data
}
\description{
Calculate the onset value of the mid summer drought from a RasterBrick or a Time Series.

The input must be in the form of daily data, with the first data point being January 1st of a respective year.

If x is a RasterBrick, then the output is a RasterBrick with a data point for each year. If you would like to find a statistical value of the new rasterbrick, it is recommended to use

r <- raster::calc(x, mean)

If the index vector from lpdates is not saved as 'dated', a function will have to be created to use it in raster::calc.

onset1<-function(x){
  onset(x,ind=vector)
}
}
\examples{
# using RasterBricks
r<-raster::calc(x, onset)

# using TimeSeries
r<-onset(x)
}
\seealso{
\code{\link{raster}}
}
