% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{dur}
\alias{dur}
\title{Duration Value}
\usage{
dur(x, date1="1950-05-01", date2="1950-07-15", date3="1950-07-16", date4="1950-09-30")
}
\arguments{
\item{x}{RasterBrick or TimeSeries}

\item{date1}{Start index to find first onset date}

\item{date2}{End index to find first onset date}

\item{date3}{Start index to find second onset date}

\item{date4}{End index to find second onset date}
}
\value{
RasterBrick or TimeSeries of Yearly data
}
\description{
Calculate the duration of the mid summer drought from a RasterBrick or a Time Series.

The input must be in the form of daily data, with the first data point being January 1st of a respective year.

If x is a RasterBrick, then the output is a RasterBrick with a data point for each year. If you would like to find a statistical value of the new rasterbrick, it is recommended to use

r <- raster::calc(x, mean)
}
\examples{
# using RasterBricks
r<-raster::calc(x, dur)

# using TimeSeries
r<-dur(x)
}
\seealso{
\code{\link{raster}}
}
